Added typecast to prevent compiler warning

+++ AccurateRip/AccurateRipQueryOperation.m
@@ -132,7 +132,7 @@
 		[accurateRipResponseData getBytes:&arFreeDBID range:NSMakeRange(pressingDataOffset + 9, 4)];
 		arFreeDBID = OSSwapLittleToHostInt32(arFreeDBID);
 		
-		if(arTrackCount != orderedTracks.count || arDiscID1 != accurateRipID1 || arDiscID2 != accurateRipID2 || arFreeDBID != compactDisc.freeDBDiscID) {
+		if(arTrackCount != orderedTracks.count || arDiscID1 != accurateRipID1 || arDiscID2 != accurateRipID2 || (NSUInteger)arFreeDBID != compactDisc.freeDBDiscID) {
 			
 #if DEBUG
 			NSLog(@"AccurateRip track count or disc IDs don't match.");
